pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "invoice-intelligence-platform"
version = "0.0.1"
description = "GST Invoice Intelligence Platform — research prototype"
requires-python = ">=3.10"
dependencies = [
    "rapidocr-onnxruntime>=1.3.0",
    "pydantic>=2.0",
    "numpy>=1.24",
    "pillow>=10.1",
    "datasets>=2.14",
    "rapidfuzz>=3.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]
tesseract = ["pytesseract>=0.3"]
paddleocr = ["paddleocr>=3.0"]
doctr = ["python-doctr[torch]>=0.7"]

[tool.hatch.build.targets.wheel]
packages = ["components", "data_sources", "tools", "backend"]

[tool.pytest.ini_options]
testpaths = ["components", "data_sources", "backend"]
python_files = ["test_*.py"]
markers = [
    "ocr_heavy: slow tests that load real OCR models (~1-2s each); deselect with '-m \"not ocr_heavy\"'",
    "dataset_heavy: tests that download or enumerate real datasets; deselect with '-m \"not dataset_heavy\"'",
]
